requirements.txt
Flask
flask-cors
Pillow
pytesseract
requests